feat: restructure preload
